+++ src/types/pageTypes.ts
@@ -1,0 +1,1 @@
+export type RCPage = "search" | "shep" | "account" | ""
